requirements.txt
alembic==1.13.1
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.51.0
greenlet==3.0.3
greenletio==0.11.0
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
JPype1 @ file:///C:/venvs/co2/files/JPype1-1.5.0-cp312-cp312-win_amd64.whl#sha256=9aafc00b00bf8c1b624081e5d4ab87f7752e6c7ee6a141cfc332250b05c6d42f
kiwisolver==1.4.5
konlpy==0.6.0
lxml==5.2.2
Mako==1.3.3
MarkupSafe==2.1.5
matplotlib==3.8.4
mod_wsgi==5.0.0
mplfinance==0.12.10b0
numpy==1.26.4
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
threadpoolctl==3.5.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.2
wordcloud==1.9.3
wsproto==1.2.0
WTForms==3.1.2
xlrd==2.0.1
